[SW-34] fix: 승하차 통계 응답 타입을 문자열 기준으로 보정

+++ domain/src/main/java/com/example/domain/model/DetailStationInfo.kt
@@ -4,6 +4,6 @@ data class DetailStationInfo(
     val usageDate: String,
     val lineName: String,
     val stationName: String,
-    val getOnCount: Double,
-    val getOffCount: Double,
+    val getOnCount: Long,
+    val getOffCount: Long,
 )
